Move and replace finish screen by google import screen
