+++ app/views/onboarding/Onboarding5.js
@@ -4,248 +4,45 @@ import {
   ImageBackground,
   StatusBar,
   StyleSheet,
-  Text,
   View,
 } from 'react-native';
-import {
-  PERMISSIONS,
-  RESULTS,
-  check,
-  checkNotifications,
-  request,
-  requestNotifications,
-} from 'react-native-permissions';
-import { SvgXml } from 'react-native-svg';
 
 import BackgroundImage from './../../assets/images/launchScreenBackground.png';
-import { isPlatformiOS } from './../../Util';
-import IconDenied from '../../assets/svgs/permissionDenied';
-import IconGranted from '../../assets/svgs/permissionGranted';
-import IconUnknown from '../../assets/svgs/permissionUnknown';
 import ButtonWrapper from '../../components/ButtonWrapper';
+import { Typography } from '../../components/Typography';
 import Colors from '../../constants/colors';
 import fontFamily from '../../constants/fonts';
-import { PARTICIPATE } from '../../constants/storage';
-import { SetStoreData } from '../../helpers/General';
+import { LOCATION_DATA } from '../../constants/storage';
+import { GetStoreData } from '../../helpers/General';
 import languages from '../../locales/languages';
-import { Typography } from '../../components/Typography';
 
 const width = Dimensions.get('window').width;
-
-const PermissionStatusEnum = {
-  UNKNOWN: 0,
-  GRANTED: 1,
-  DENIED: 2,
-};
-
-const PermissionDescription = ({ title, status, ...props }) => {
-  let icon;
-  switch (status) {
-    case PermissionStatusEnum.UNKNOWN:
-      icon = IconUnknown;
-      break;
-    case PermissionStatusEnum.GRANTED:
-      icon = IconGranted;
-      break;
-    case PermissionStatusEnum.DENIED:
-      icon = IconDenied;
-      break;
-  }
-  return (
-    <View style={styles.permissionContainer}>
-      <Typography style={styles.permissionTitle}>{title}</Typography>
-      <SvgXml style={styles.permissionIcon} xml={icon} width={30} height={30} />
-    </View>
-  );
-};
 
 class Onboarding extends Component {
   constructor(props) {
     super(props);
+
+    let storedLocations = GetStoreData(LOCATION_DATA, false);
+    storedLocations = Array.isArray(storedLocations) ? storedLocations : [];
     this.state = {
-      notificationPermission: PermissionStatusEnum.UNKNOWN,
-      locationPermission: PermissionStatusEnum.UNKNOWN,
+      locationData: storedLocations,
     };
-    this.checkLocationStatus();
-    this.checkNotificationStatus();
   }
 
-  isLocationChecked() {
-    return this.state.locationPermission !== PermissionStatusEnum.UNKNOWN;
-  }
-
-  isNotificationChecked() {
-    return this.state.notificationPermission !== PermissionStatusEnum.UNKNOWN;
-  }
-
-  checkLocationStatus() {
-    // NEED TO TEST ON ANNDROID
-    let locationPermission;
-    if (isPlatformiOS()) {
-      locationPermission = PERMISSIONS.IOS.LOCATION_ALWAYS;
-    } else {
-      locationPermission = PERMISSIONS.ANDROID.ACCESS_FINE_LOCATION;
-    }
-    check(locationPermission)
-      .then(result => {
-        switch (result) {
-          case RESULTS.GRANTED:
-            this.setState({
-              locationPermission: PermissionStatusEnum.GRANTED,
-            });
-            break;
-          case RESULTS.UNAVAILABLE:
-          case RESULTS.BLOCKED:
-            this.setState({
-              locationPermission: PermissionStatusEnum.DENIED,
-            });
-            break;
-        }
-      })
-      .catch(error => {
-        console.log('error checking location: ' + error);
-      });
-  }
-
-  checkNotificationStatus() {
-    checkNotifications().then(({ status }) => {
-      switch (status) {
-        case RESULTS.GRANTED:
-          this.setState({
-            notificationPermission: PermissionStatusEnum.GRANTED,
-          });
-          break;
-        case RESULTS.UNAVAILABLE:
-        case RESULTS.BLOCKED:
-          this.setState({
-            notificationPermission: PermissionStatusEnum.DENIED,
-          });
-          break;
-      }
-    });
-  }
-
-  requestLocation() {
-    // NEED TO TEST ON ANNDROID
-    let locationPermission;
-    if (isPlatformiOS()) {
-      locationPermission = PERMISSIONS.IOS.LOCATION_ALWAYS;
-    } else {
-      locationPermission = PERMISSIONS.ANDROID.ACCESS_FINE_LOCATION;
-    }
-    request(locationPermission).then(result => {
-      switch (result) {
-        case RESULTS.GRANTED:
-          console.log('Location granted');
-          this.setState({
-            locationPermission: PermissionStatusEnum.GRANTED,
-          });
-          break;
-        case RESULTS.UNAVAILABLE:
-        case RESULTS.BLOCKED:
-          this.setState({
-            locationPermission: PermissionStatusEnum.DENIED,
-          });
-          break;
-      }
-    });
-  }
-
-  requestNotification() {
-    requestNotifications(['alert', 'badge', 'sound']).then(({ status }) => {
-      switch (status) {
-        case RESULTS.GRANTED:
-          this.setState({
-            notificationPermission: PermissionStatusEnum.GRANTED,
-          });
-          break;
-        case RESULTS.UNAVAILABLE:
-        case RESULTS.BLOCKED:
-          this.setState({
-            notificationPermission: PermissionStatusEnum.DENIED,
-          });
-          break;
-      }
-    });
-  }
-
-  buttonPressed() {
-    if (!this.isLocationChecked()) {
-      this.requestLocation();
-    } else if (!this.isNotificationChecked()) {
-      this.requestNotification();
-    } else {
-      SetStoreData(PARTICIPATE, 'true'); // replaces "start" button
-      SetStoreData('ONBOARDING_DONE', true);
-      this.props.navigation.replace('LocationTrackingScreen');
-    }
-  }
-
-  getTitleText() {
-    if (!this.isLocationChecked()) {
-      return languages.t('label.launch_location_header');
-    } else if (!this.isNotificationChecked()) {
-      return languages.t('label.launch_notif_header');
-    } else {
-      return languages.t('label.launch_done_header');
-    }
-  }
-
-  getTitleTextView() {
-    if (!this.isLocationChecked() || !this.isNotificationChecked()) {
-      return <Typography style={styles.headerText}>{this.getTitleText()}</Typography>;
-    } else {
-      return <Typography style={styles.bigHeaderText}>{this.getTitleText()}</Typography>;
-    }
-  }
-
-  getSubtitleText() {
-    if (!this.isLocationChecked()) {
-      return languages.t('label.launch_location_subheader');
-    } else if (!this.isNotificationChecked()) {
-      return languages.t('label.launch_notif_subheader');
-    } else {
-      return languages.t('label.launch_done_subheader');
-    }
-  }
-
-  getLocationPermission() {
+  hasLocationData = () => {
     return (
-      <>
-        <View style={styles.divider}></View>
-        <PermissionDescription
-          title={languages.t('label.launch_location_access')}
-          status={this.state.locationPermission}
-        />
-        <View style={styles.divider}></View>
-      </>
+      Array.isArray(this.state.locationData) &&
+      this.state.locationData.length > 0
     );
-  }
+  };
 
-  getNotificationsPermissionIfIOS() {
-    if (isPlatformiOS()) {
-      return (
-        <>
-          <PermissionDescription
-            title={languages.t('label.launch_notification_access')}
-            status={this.state.notificationPermission}
-          />
-          <View style={styles.divider}></View>
-        </>
-      );
-    }
-    return;
-  }
+  onGoToSettingsPress = () => {
+    this.props.navigation.navigate('ImportScreen');
+  };
 
-  getButtonText() {
-    if (!this.isLocationChecked()) {
-      return languages.t('label.launch_enable_location');
-    } else if (!this.isNotificationChecked()) {
-      return languages.t('label.launch_enable_notif');
-    } else {
-      return languages.t('label.launch_finish_set_up');
-    }
-  }
+  onButtonPress = () => {
+    this.props.navigation.replace('Onboarding6');
+  };
 
   render() {
     return (
@@ -253,23 +50,37 @@ class Onboarding extends Component {
         <StatusBar
           barStyle='light-content'
           backgroundColor='transparent'
-          translucent={true}
+          translucent
         />
-
         <View style={styles.mainContainer}>
           <View style={styles.contentContainer}>
-            {this.getTitleTextView()}
-            <Typography style={styles.subheaderText}>{this.getSubtitleText()}</Typography>
+            <Typography style={styles.headerText}>
+              {languages.t('label.maps_import_title')}
+            </Typography>
+            <Typography style={styles.subheaderText}>
+              {languages.t('label.maps_import_text')}
+            </Typography>
             <View style={styles.statusContainer}>
-              {this.getLocationPermission()}
-              {this.getNotificationsPermissionIfIOS()}
-              <View style={styles.spacer}></View>
+              <View style={styles.spacer} />
             </View>
           </View>
           <View style={styles.footerContainer}>
+            {!this.hasLocationData() && (
+              <ButtonWrapper
+                title={languages.t('label.maps_import_button_text')}
+                onPress={this.onGoToSettingsPress}
+                buttonColor={Colors.VIOLET}
+                bgColor={Colors.WHITE}
+              />
+            )}
+            <View style={styles.spacer} />
             <ButtonWrapper
-              title={this.getButtonText()}
-              onPress={this.buttonPressed.bind(this)}
+              title={
+                this.hasLocationData()
+                  ? languages.t('label.launch_next')
+                  : 'Skip'
+              }
+              onPress={this.onButtonPress}
               buttonColor={Colors.VIOLET}
               bgColor={Colors.WHITE}
             />
@@ -296,14 +107,6 @@ const styles = StyleSheet.create({
     justifyContent: 'center',
     alignSelf: 'center',
   },
-  bigHeaderText: {
-    color: Colors.WHITE,
-    fontSize: 52,
-    lineHeight: 48.5,
-    paddingTop: 52 - 48.5, // lineHeight hack
-    width: width * 0.7,
-    fontFamily: fontFamily.primaryMedium,
-  },
   headerText: {
     color: Colors.WHITE,
     fontSize: 26,
@@ -320,11 +123,6 @@ const styles = StyleSheet.create({
   statusContainer: {
     marginTop: '5%',
   },
-  divider: {
-    backgroundColor: Colors.DIVIDER,
-    height: 1,
-    marginVertical: '3%',
-  },
   spacer: {
     marginVertical: '5%',
   },
@@ -332,19 +130,6 @@ const styles = StyleSheet.create({
     position: 'absolute',
     bottom: 0,
     marginBottom: '10%',
-    alignSelf: 'center',
-  },
-  permissionContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-  },
-  permissionTitle: {
-    color: Colors.WHITE,
-    fontSize: 16,
-    alignSelf: 'center',
-    fontFamily: fontFamily.primaryRegular,
-  },
-  permissionIcon: {
     alignSelf: 'center',
   },
 });
